Finished day 16, QA is next. See CHANGE.LOG for details.

+++ code/handmade.cpp
@@ -42,8 +42,8 @@ RenderWeirdGradient(game_offscreen_buffer *Buffer, int BlueOffset, int GreenOffs
 			Pixel in Memory: BB GG RR xx
 			0x xxRRGGBB
 			*/
-			uint8 Blue = (X + BlueOffset);
-			uint8 Green = (Y + GreenOffset);
+			uint8 Blue = (uint8)(X + BlueOffset);
+			uint8 Green = (uint8)(Y + GreenOffset);
 
 			*Pixel++ = ((Green << 8) | Blue);
 		}
@@ -80,8 +80,8 @@ GameUpdateAndRender(game_memory *Memory, game_input *Input, game_offscreen_buffe
     if(Input0->IsAnalog)
     {
         /// NOTE(Dennis): Use analog movement tuning
-        GameState->BlueOffset += (int)4.0f*(Input0->EndX);
-        GameState->ToneHz = 512 + (int)(128.0f*(Input0->EndY));
+        GameState->BlueOffset += (int)(4.0f*Input0->EndX);
+        GameState->ToneHz = 512 + (int)(128.0f*Input0->EndY);
     }
     else
     {
